report-job/requirements.txt
azure-storage-blob==12.19.0
reportlab==4.0.7
